Removed the need to inject global variables for loading plugins.
Plugins are no longer dependant on an injected global method/variable for loading. We now create a small function inside the sandbox that handles loading of the plugin and returning it's running instance. This way we can call the function from outside the sandbox to retrieve a proxied instance of the plugin. This commit also enables access to the 'events' built-in module inside the sandbox.